refactor: 重构代码结构，将A股相关模块移动到ChinaTrade命名空间
将A股相关的接口、模型、策略等代码移动到ChinaTrade命名空间下，以提高代码的可维护性和组织结构。同时更新了配置文件中的环境设置。
